Fix windows libtbx-build compilation for modern C++ (#605)
libtbx-build built-in paths were overriding the system-provided standard
library with custom versions of - at least - stdint.h. This caused failures
when trying to use C++14, but possibly some features of C++11 also.

This mainly seems to have been an issue with the MS-provided standard library
on windows, but we probably don't want this to happen on any platform.